Two rollup configurations, speeds up build/test cycles
